:recycle: feat: Post 디렉토리 삭제 및 rename
