+++ src/main/java/com/miniproject/demo/domain/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
 package com.miniproject.demo.domain.post.repository;
 
-import com.miniproject.demo.domain.post.domain.Post;
+import com.miniproject.demo.domain.post.entity.Post;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
