+++ src/main/java/com/miniproject/demo/domain/post/service/PostService.java
@@ -1,6 +1,6 @@
 package com.miniproject.demo.domain.post.service;
 
-import com.miniproject.demo.domain.post.domain.Post;
+import com.miniproject.demo.domain.post.entity.Post;
 import com.miniproject.demo.domain.post.dto.PostRequestDTO;
 
 import java.util.List;
